Refactor to Use TableViewController
- Refactor ViewController into Table View Controller
- Refactor Searchbar to Search Controller
