+++ 20240328-DanielSpady-Assement-Country/Features/Country/List/Controller/Extension/TableViewController+DataSource.swift
@@ -1,15 +1,15 @@
 //
-//  ViewController+TableView.swift
+//  TableViewController+DataSource.swift
 //  20240328-DanielSpady-Assement-Country
 //
-//  Created by Daniel Spady on 2024-03-28.
+//  Created by Daniel Spady on 2024-03-31.
 //
 
 import UIKit
 
-extension ViewController: UITableViewDelegate, UITableViewDataSource {
+extension TableViewController {
     
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell = tableView.dequeueReusableCell(withIdentifier: "tableViewCell", for: indexPath) as! CountryTableViewCell
         let viewModel = filteredCountryList[indexPath.row]
         cell.nameLabel.text = "\(viewModel.name ?? ""), \(viewModel.region ?? "")"
@@ -17,12 +17,12 @@ extension ViewController: UITableViewDelegate, UITableViewDataSource {
         cell.capitalLabel.text = viewModel.capital
         return cell
     }
-
-    func numberOfSectionsInTableView(tableView: UITableView) -> Int {
+    
+    override func numberOfSections(in tableView: UITableView) -> Int {
         return 1
     }
 
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         return filteredCountryList.count
     }
 }
