+++ 20240328-DanielSpady-Assement-Country/Features/Country/List/Controller/TableViewController.swift
@@ -1,0 +1,29 @@
+//
+//  TableViewController.swift
+//  20240328-DanielSpady-Assement-Country
+//
+//  Created by Daniel Spady on 2024-03-31.
+//
+
+import UIKit
+
+class TableViewController: BaseTableViewController {
+    
+    @IBOutlet weak var errorView: UIView!
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Do any additional setup after loading the view.
+        
+        tableView.delegate = self
+        tableView.dataSource = self
+        
+        let searchController = UISearchController(searchResultsController: nil)
+        searchController.searchResultsUpdater = self
+        searchController.obscuresBackgroundDuringPresentation = false
+        searchController.searchBar.placeholder = "Search Country"
+        navigationItem.searchController = searchController
+        
+        self.loadData(for: tableView, errorView: errorView)
+    }
+}
